# Improve performance around Axis(3) limits