--- a/src/makielayout/blocks/axis3d.jl
+++ b/src/makielayout/blocks/axis3d.jl
@@ -282,7 +282,7 @@ function Makie.plot!(
 
     plot = Makie.plot!(ax.scene, P, allattrs, args...)
 
-    if ax.reset_limits_on_plot[]
+    if is_open_or_any_parent(ax.scene)
         reset_limits!(ax)
     end
     plot
@@ -291,6 +291,11 @@ end
 function Makie.plot!(P::Makie.PlotFunc, ax::Axis3, args...; kw_attributes...)
     attributes = Makie.Attributes(kw_attributes)
     Makie.plot!(ax, P, attributes, args...)
+end
+
+function update_state_before_display!(ax::Axis3)
+    reset_limits!(ax)
+    return
 end
 
 function autolimits!(ax::Axis3)