Seems like performance slows quite a bit if many plots are added in sequence to an Axis. Because each plot triggers a `resize_limits!`, first of all the limits have to be recomputed for all previous plots when a new one is added, and this also triggers tick changes, which trigger layout changes, etc. So all in all this can be quite slow. This PR adds an option to disable the `reset_limits!` call, which should mainly be interesting for workflows where all plots are added in one go, without a person looking at the progress. So for example this should probably be used in AlgebraOfGraphics @piever .

Also, I noticed dynamic dispatch in the code path that computes the limits in the first place and tried to fix that. I'm not sure if the logic is perfectly right with regards to NaNs, that behavior can be a bit tricky to pin down. For example, what if there's a NaN only in one dimension of a point, do we skip that whole point or use the remaining components. This PR does the latter.

Here's some examples how the PR improves performance:

## Many plots in sequence
Before: 4.999897 seconds (206.07 M allocations: 6.235 GiB, 14.02% gc time)
After: 0.255500 seconds (2.88 M allocations: 151.533 MiB)
```julia
f = Figure()
ax = Axis(f[1, 1])
ax.reset_limits_on_plot = false # new feature
@time for i in 1:500
    lines!(ax, randn(1000) .+ 5i)
end
ax.reset_limits_on_plot = true # new feature
reset_limits!(ax)
f
```

## Single plot
Before:  0.176131 seconds (10.00 M allocations: 305.179 MiB)
After: 0.012864 seconds (92 allocations: 4.203 KiB)
```julia
f = Figure()
ax = Axis(f[1, 1])
lines!(cumsum(randn(1_000_000)))
@time reset_limits!(ax)
f
```

## Type of change

Delete options that do not apply:

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)

## Checklist

- [ ] Added an entry in NEWS.md (for new features and breaking changes)
- [ ] Added or changed relevant sections in the documentation
- [ ] Added unit tests for new algorithms, conversion methods, etc.
- [ ] Added reference image tests for new plotting functions, recipes, visual options, etc.