@SimonDanisch 2022-07-05
I still think we should only update limits on display, before we add the workaround of disabling limit calculation to all kind of 3rd party libraries... This could be quite the easy change, no?