@SimonDanisch 2022-07-05
Adding something like
```julia
update_state_before_display!(figlike) # or some better name
```
Here: https://github.com/JuliaPlots/Makie.jl/blob/master/src/display.jl#L62=
And overload that function appropriately for figure, scene, axis etc, so that it gets called for all axes in a figure.