@Moelf 2022-07-05
I think compute on display is reasonable, `display` is already non-pure (I/O).

One problem I can think of is: do people use reflection (to read axis limits) before display in order to do something? I hope the answer is no