@jkrumbiegel 2022-07-05
In that case they could always call `reset_limits!` themselves I guess?