@ffreyer 2022-07-05
Wouldn't update on display be an issue when plotting in the repl? I.e. starting with `fig = Figure()` in the REPL you'd display immediately and need another display or `reset_limit!` at the end to update limits.

Another option to throttle these kind of updates would be to generalize `screen.render_tick` from GLMakie to all backends and use it as a trigger. That would reduce things to one update per frame but it would require some bookkeeping to decide what needs to be updates.