@jkrumbiegel 2022-07-07
Could we do it such that in the `plot!` call we check if any parent `Scene` of the `Axis` is currently displayed? If yes, we reset limits. If not, we don't (most programmatic cases).

And additionally the `update_state_before_display!` thing that Simon suggested.