@jkrumbiegel 2022-07-08
I've changed the implementation to `update_state_before_display!` and reset if the Axis scene or any of its parents are open.

However, there's still the problem of all the functions that can save a scene. `save`, `record`, `colorbuffer`, `VideoStream`, are there more? Do all of them need a separate `update_state_before_display!` inserted?