@SimonDanisch 2022-07-08
Most of them should use backend_display / backend_show under the hood... I can take a look if you want