@jkrumbiegel 2022-07-09
Tests pass now, question is if all the places to call `update_state_before_display!` make sense.