@SimonDanisch 2022-07-20
Guess it would be good to check the docs before merging this ;)